feat: added template function for fetching an event
